Some minor refactorings to keep isolation in QueryRunner and various other places
- Also move methods belonging to ModifySpaceEvent handling into "ModifySpace" query-runner
