+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/QueryRunner.java
@@ -38,29 +38,29 @@ public abstract class QueryRunner<E extends Event, R extends XyzResponse> implem
   protected static final String SCHEMA = "schema";
   protected static final String TABLE = "table";
 
-  protected Event event;
   private final SQLQuery query;
   private boolean useReadReplica;
   protected DatabaseHandler dbHandler;
 
   public QueryRunner(E event, DatabaseHandler dbHandler) throws SQLException {
-    this.event = event;
     this.dbHandler = dbHandler;
     query = buildQuery(event);
   }
 
   public R run() throws SQLException {
-    query.replaceFragments();
-    query.replaceVars();
-    query.replaceNamedParameters();
+    prepareQuery();
     return dbHandler.executeQueryWithRetry(query, this, useReadReplica);
   }
 
-  public int write() throws SQLException {
+  public void write() throws SQLException {
+    prepareQuery();
+    dbHandler.executeUpdateWithRetry(query);
+  }
+
+  private void prepareQuery() {
     query.replaceFragments();
     query.replaceVars();
     query.replaceNamedParameters();
-    return dbHandler.executeUpdateWithRetry(query);
   }
 
   protected abstract SQLQuery buildQuery(E event) throws SQLException;
@@ -78,5 +78,9 @@ public abstract class QueryRunner<E extends Event, R extends XyzResponse> implem
 
   protected String getSchema() {
     return dbHandler.config.getDatabaseSettings().getSchema();
+  }
+
+  protected String getDefaultTable(E event) {
+    return dbHandler.config.readTableFromEvent(event);
   }
 }
