+++ xyz-psql-connector/src/main/java/com/here/xyz/psql/query/SearchForFeatures.java
@@ -43,7 +43,7 @@ public class SearchForFeatures<E extends SearchForFeaturesEvent> extends Extende
   protected boolean hasSearch;
 
   public SearchForFeatures(E event, DatabaseHandler dbHandler) throws SQLException {
-    super(event, dbHandler,false);
+    super(event, dbHandler);
   }
 
   @Override
